From Simulation on 10/12/25
Ran this simulation with the 100K Rao Nozzle Mesh on 10/12/25 and got some very interesting results towards the end. This particular video is the mach contour for the engine. Going to keep this file on my main desktop to plot some things in the chamber section to verify alignment with CEA values per thesis goals.